+++ src/erp.jl
@@ -229,13 +229,20 @@ Spherical linear interpolation between two EulerRodriguesParameters.
 The parameter `f` can range from 0 to 1 to specify where along the sphere to interpolate.
 When f = 0, the result is `a` and when f = 1, the result is `b`.
 
-Setting `shortest_path` to false will allow the interpolation to take the "long way around".
+Setting `shortest_path` to false preserves the path encoded by the signs of the supplied
+Euler parameters. This can take the "long way around" when the supplied parameters are on
+the long arc; use `other(a)` or `other(b)` to request that path explicitly.
 
 References
 - https://en.wikipedia.org/wiki/Slerp
 - https://blog.magnum.graphics/backstage/the-unnecessarily-short-ways-to-do-a-quaternion-slerp/
 """
-function interpolate(ep_start::EulerRodriguesParameters{T}, ep_end::EulerRodriguesParameters{T}, f; shortest_path::Bool = true) where {T}
+function interpolate(
+    ep_start::EulerRodriguesParameters{T},
+    ep_end::EulerRodriguesParameters{T},
+    f;
+    shortest_path::Bool = true,
+) where {T}
 
     if f < 0 || f > 1
         throw(DomainError(f, "f was $f but must be in the range [0, 1]."))
@@ -249,28 +256,35 @@ function interpolate(ep_start::EulerRodriguesParameters{T}, ep_end::EulerRodrigu
 
     # This is the intuitive implementation, but what's below is faster.
     #
-    #   aa = erp2aa(smallest(difference(ep_end, ep_start)))
+    #   d = difference(ep_end, ep_start)
+    #   if shortest_path
+    #       d = smallest(d)
+    #   end
+    #   aa = erp2aa(d)
     #   return normalize(compose(aa2erp(AxisAngle(aa.axis, aa.angle * f)), ep_start))
-    #
 
     d = dot(ep_start, ep_end)
-    if d <= -one(T) && !shortest_path
-        return ep_start
-    end
-
-    if shortest_path
+    if shortest_path && d < zero(T)
         # Flip `ep_start` if there is a shorter path.
-        ep_startp = d >= 0 ? ep_start : other(ep_start)
+        ep_startp = other(ep_start)
+        d = -d
     else
         ep_startp = ep_start
     end
+    if d <= -one(T)
+        return ep_start
+    end
 
-    θ = acos(clamp(abs(d), zero(T), one(T)))
+    # The angle here will be positive by virtue of the acos.
+    θ = acos(clamp(d, -one(T), one(T)))
+
+    # For very small angles from start to end, we skip the trig (and hence avoid dividing
+    # by 0).
     if θ <= sqrt(eps(float(T)))
         return normalize((1 - f) * ep_startp + f * ep_end)
     end
 
-    # If -1 < cos(θ) < 1, then sin(θ) should be > 0.
+    # If -1 < cos(θ) < 1, then sin(θ) should be nonzero.
     c1 = sin((1 - f) * θ) / sin(θ)
     c2 = sin(f * θ) / sin(θ)
     return normalize(c1 * ep_startp + c2 * ep_end)