+++ changelog.md
@@ -1,5 +1,9 @@
 # Orientations.jl Change Log
 
+## Unreleased
+
+* Fixed ERP interpolation when `shortest_path = false`.
+
 ## v1.0
 
 This is the initial release.